+++ wifi/services/wifi_standard/wifi_framework/wifi_toolkit/config/wifi_settings.cpp
@@ -471,7 +471,7 @@ int WifiSettings::RemoveExcessDeviceConfigs(std::vector<WifiDeviceConfig> &confi
     if (maxNumConfigs < 0) {
         return 1;
     }
-    int numExcessNetworks = configs.size() - maxNumConfigs;
+    int numExcessNetworks = static_cast<int>(configs.size()) - maxNumConfigs;
     if (numExcessNetworks <= 0) {
         return 1;
     }
